Backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow>=10.0.0
reportlab>=4.0.0
pydantic>=2.0.0
